+++ src/components/schema-table.js
@@ -3,6 +3,7 @@ import { marked } from 'marked';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import FontStyles from '../styles/font-styles.js';
 import SchemaStyles from '../styles/schema-styles';
+import KeyFrameStyles from '../styles/key-frame-styles.js';
 
 const tablePadding = 16;
 const firstColumnInitialPadding = tablePadding * 2;
@@ -29,6 +30,7 @@ export default class SchemaTable extends LitElement {
   static finalizeStyles() {
     return [
       FontStyles,
+      KeyFrameStyles,
       SchemaStyles,
       css`
       .table {
@@ -74,12 +76,13 @@ export default class SchemaTable extends LitElement {
         background-clip: border-box;
       }
       .tr + .object-body {
-        max-height: 5000px;
-        transition: max-height 1.2s ease-in-out;
         overflow: hidden;
+      } 
+      .tr:not(.collapsed) + .object-body {
+        animation: linear 0.2s expand-height;
       }
       .tr.collapsed + .object-body {
-        transition: max-height 1.2s ease-in-out -1.0s;
+        animation: linear 0.2s collapse-height;
         max-height: 0;
       }
       .obj-toggle {
@@ -112,7 +115,7 @@ export default class SchemaTable extends LitElement {
     `;
   }
 
-  generateTree(data, dataType = 'object', key = '', description = '', schemaLevel = 0, indentLevel = 0) {
+  generateTree(data, dataType = 'object', key = '', title = '', description = '', schemaLevel = 0, indentLevel = 0) {
     const newSchemaLevel = data['::type'] && data['::type'].startsWith('xxx-of') ? schemaLevel : (schemaLevel + 1);
     const newIndentLevel = dataType === 'xxx-of-option' || data['::type'] === 'xxx-of-option' || key.startsWith('::OPTION') ? indentLevel : (indentLevel + 1);
     // 16px space indentation at each level, start the first one at 32px to align with the field hr key row object
@@ -164,7 +167,7 @@ export default class SchemaTable extends LitElement {
         return undefined;
       }
       
-      const displayLine = [description].filter(v => v).join(' ');
+      const displayLine = [title && `**${title}:**`, description].filter(v => v).join(' ');
       return html`
         ${newSchemaLevel >= 0 && key
           ? html`
@@ -194,13 +197,13 @@ export default class SchemaTable extends LitElement {
           `
         }
         <div class='object-body'>
-        ${Array.isArray(data) && data[0] ? html`${this.generateTree(data[0], 'xxx-of-option', '::ARRAY~OF', '', newSchemaLevel, newIndentLevel)}`
+        ${Array.isArray(data) && data[0] ? html`${this.generateTree(data[0], 'xxx-of-option', '::ARRAY~OF', data[0]['::title'], data[0]['::description'], newSchemaLevel, newIndentLevel)}`
             : html`
               ${Object.keys(data).map((dataKey) =>
                 !['::title', '::description', '::type', '::props', '::deprecated', '::array-type', '::dataTypeLabel', '::flags'].includes(dataKey)
                 || data[dataKey]['::type'] === 'array' && data[dataKey]['::type'] === 'object'
                 ? html`${this.generateTree(data[dataKey]['::type'] === 'array' ? data[dataKey]['::props'] : data[dataKey],
-                      data[dataKey]['::type'], dataKey, data[dataKey]['::description'], newSchemaLevel, newIndentLevel)}`
+                      data[dataKey]['::type'], dataKey, data[dataKey]['::title'], data[dataKey]['::description'], newSchemaLevel, newIndentLevel)}`
                 : ''
               )}`
           }
@@ -231,9 +234,8 @@ export default class SchemaTable extends LitElement {
           <div class="attributes ${cssType}" style="font-family: var(--font-mono);" title="${readOrWriteOnly === '🆁' && 'Read only attribute' || readOrWriteOnly === '🆆' && 'Write only attribute' || ''}">${readOrWriteOnly}</div>
         </div>
         <div class='td key-descr'>
-          ${dataType === 'array' ? html`<span class="m-markdown-small">${unsafeHTML(marked(description))}</span>` : ''}
           <span class="m-markdown-small" style="vertical-align: middle;">
-            ${unsafeHTML(marked(`${dataType === 'array' && description || `${schemaTitle ? `**${schemaTitle}:**` : ''} ${schemaDescription}` || ''}`))}
+            ${unsafeHTML(marked(`${`${(schemaTitle || title) ? `**${schemaTitle || title}:**` : ''} ${schemaDescription || description}` || ''}`))}
           </span>
           ${this.schemaDescriptionExpanded ? html`
             ${constraint ? html`<div style='display:inline-block; line-break: anywhere; margin-right:8px;'><span class='bold-text'>Constraints: </span>${constraint}</div><br>` : ''}