+++ CHANGELOG.md
@@ -11,6 +11,13 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Prevent bad bodies from being passed to GET and HEAD, GET and HEAD do not take bodies in requests
 * Don't copy the 8 spaces <pre> formatted response from JSON requests #138.
 * Fix initial padding for non-object based field properties, by setting the min at 32px (or 2 * the shift which is 16px)
+* Ensure UI recordings do not include curl command which might include security tokens
+* Fix table layout so that the path parameters table has a wider input area.
+* Don't filter out tags that have no paths, no-path tags can be used for documentation.
+* Fix pattern truncation to display full pattern on hover.
+* Center description in request forms when no constraints are specified.
+* Fix nested one-of descriptions.
+* Fix example display for request path + query parameters.
 
 ## 0.11 ##
 * Fix `allOf` for response schema. #119